+++ src/Component/tests/Symfony/Routing/RedirectHandlerTest.php
